fix: getting an empty list of machines
When an empty list is returned (aka, the first time it's ever executed),
trying to inspect the machines won't work as a simple
   podman machine inspect
is run.
By default, the command tries to inspect the default machine,
which also doesn't exist --> returncode != 0

When we get an empty list, we simply don't inspect anymore.